# [Core][MultiModalHasher] Don't convert memoryviews to bytes during hashing