--- a/vllm/multimodal/hasher.py
+++ b/vllm/multimodal/hasher.py
@@ -23,22 +23,21 @@ class MultiModalHasher:
     def serialize_item(cls, obj: object) -> Iterable[Union[bytes, memoryview]]:
         # Simple cases
         if isinstance(obj, (bytes, memoryview)):
-            yield obj
-        elif isinstance(obj, str):
-            yield obj.encode("utf-8")
-        elif isinstance(obj, (int, float)):
-            yield np.array(obj).tobytes()
+            return (obj, )
+        if isinstance(obj, str):
+            return (obj.encode("utf-8"), )
+        if isinstance(obj, (int, float)):
+            return (np.array(obj).tobytes(), )
 
-        elif isinstance(obj, Image.Image):
+        if isinstance(obj, Image.Image):
             exif = obj.getexif()
             if Image.ExifTags.Base.ImageID in exif and isinstance(
                     exif[Image.ExifTags.Base.ImageID], uuid.UUID):
                 # If the image has exif ImageID tag, use that
-                yield exif[Image.ExifTags.Base.ImageID].bytes
-            else:
-                yield from cls.iter_item_to_bytes(
-                    "image", np.asarray(convert_image_mode(obj, "RGBA")))
-        elif isinstance(obj, torch.Tensor):
+                return (exif[Image.ExifTags.Base.ImageID].bytes, )
+            return cls.iter_item_to_bytes(
+                "image", np.asarray(convert_image_mode(obj, "RGBA")))
+        if isinstance(obj, torch.Tensor):
             tensor_obj: torch.Tensor = obj.cpu()
             tensor_dtype = tensor_obj.dtype
             tensor_shape = tensor_obj.shape
@@ -50,29 +49,27 @@ class MultiModalHasher:
                 tensor_obj = tensor_obj.view(
                     (tensor_obj.numel(), )).view(torch.uint8)
 
-                yield from cls.iter_item_to_bytes(
+                return cls.iter_item_to_bytes(
                     "tensor", {
                         "original_dtype": str(tensor_dtype),
                         "original_shape": tuple(tensor_shape),
                         "data": tensor_obj.numpy(),
                     })
-            else:
-                yield from cls.iter_item_to_bytes("tensor", tensor_obj.numpy())
-        elif isinstance(obj, np.ndarray):
+            return cls.iter_item_to_bytes("tensor", tensor_obj.numpy())
+        if isinstance(obj, np.ndarray):
             # If the array is non-contiguous, we need to copy it first
             arr_data = obj.view(
                 np.uint8).data if obj.flags.c_contiguous else obj.tobytes()
-            yield from cls.iter_item_to_bytes("ndarray", {
+            return cls.iter_item_to_bytes("ndarray", {
                 "dtype": obj.dtype.str,
                 "shape": obj.shape,
                 "data": arr_data,
             })
-        else:
-            logger.warning(
-                "No serialization method found for %s. "
-                "Falling back to pickle.", type(obj))
+        logger.warning(
+            "No serialization method found for %s. "
+            "Falling back to pickle.", type(obj))
 
-            yield pickle.dumps(obj)
+        return (pickle.dumps(obj), )
 
     @classmethod
     def iter_item_to_bytes(