## Purpose

As mentioned in #22044 hashing large multimodal input can be inefficient. #19484 removed a data copy of numpy arrays by relying on memory views for c-contiguous data.

`blake3` already accepts uint8 buffers so this PR removes the need to convert everything back to bytes in `item_to_bytes`
https://github.com/vllm-project/vllm/blob/5bcc153d7bf69ef34bc5788a33f60f1792cf2861/vllm/multimodal/hasher.py#L81

Implementation wise `serialize_item` now yields either `bytes` or a `memoryview` which `blake3` directly consumes.

## Test Plan

Correctness should be covered by the existing hasher tests on CI.

The performance can be measured using:
```python
import numpy as np
from vllm.multimodal.hasher import MultiModalHasher

np.random.seed(42)
data = np.random.randn(3840, 2160, 4)

%timeit MultiModalHasher.hash_kwargs(data=data)
```

## Test Result

For a 4k image this speeds up hashing by ~30%. This is not massive, but for a lot of multimodal input this might become noticeably.

```shell
# main
168 ms ± 892 μs per loop (mean ± std. dev. of 7 runs, 10 loops each)
# This PR
120 ms ± 1.09 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```