@kexinoh 2025-09-28
This solution still has problems. In fact, there are still many parameters in the image that may affect its state, such as the tRNS state.
@lgeiger