@DarkLight1337 2025-09-28
This PR aims at addressing efficiency. The uniqueness of the hash should be same as before. If that's not the case, please open a separate issue and provide more details